# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Server API for Hamster Money Service Application",
        "title": "Hamster API",
        "contact": {
            "name": "Hamster API Support",
            "url": "https://t.me/CodeMaster482",
            "email": "grigorikovalenko@gmail.com"
        },
        "version": "1.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/user/{userID}/account/feed",
    "paths": {
...
        "/api/transaction/count": {
            "get": {
                "description": "Get User count transaction",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Get count transaction",
                "responses": {
                    "200": {
                        "description": "Show transaction count",
                        "schema": {
                            "$ref": "#/definitions/http.Response-http_TransactionCount"
                        }
                    },
                    "400": {
                        "description": "Client error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/transaction/create": {
            "post": {
                "description": "Create transaction",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Create transaction",
                "parameters": [
                    {
                        "description": "Input transactin create",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CreateTransaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create transaction",
                        "schema": {
                            "$ref": "#/definitions/http.Response-http_TransactionCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Client error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/transaction/feed": {
            "get": {
                "description": "Get User all transaction",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Get all transaction",
                "parameters": [
                    {
                        "type": "string",
                        "example": "uuid",
                        "name": "account",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "uuid",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2023-12-21T19:30:57+03:00",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "example": true,
                        "name": "income",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "example": true,
                        "name": "outcome",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2023-11-21T19:30:57+03:00",
                        "name": "start_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show transaction",
                        "schema": {
                            "$ref": "#/definitions/http.Response-http_MasTransaction"
                        }
                    },
                    "204": {
                        "description": "Show actual accounts",
                        "schema": {
                            "$ref": "#/definitions/http.Response-string"
                        }
                    },
                    "400": {
                        "description": "Client error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/transaction/update": {
            "put": {
                "description": "Put transaction",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "PUT Update",
                "parameters": [
                    {
                        "description": "Input transactin update",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.UpdTransaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update transaction",
                        "schema": {
                            "$ref": "#/definitions/http.Response-http_NilBody"
                        }
                    },
                    "400": {
                        "description": "Client error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden user",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/transaction/{transaction_id}/delete": {
            "delete": {
                "description": "Delete transaction with chosen ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Delete Transaction",
                "responses": {
                    "200": {
                        "description": "Transaction deleted",
                        "schema": {
                            "$ref": "#/definitions/http.Response-http_NilBody"
                        }
                    },
                    "400": {
                        "description": "Transaction error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "401": {
                        "description": "User unathorized",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "403": {
                        "description": "User hasn't rights",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/http.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http.CreateTransaction": {
            "type": "object",
            "properties": {
                "account_income": {
                    "description": "???",
                    "type": "string"
                },
                "account_outcome": {
                    "description": "???",
                    "type": "string"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "income": {
                    "type": "number"
                },
                "outcome": {
                    "type": "number"
                },
                "payer": {
                    "type": "string"
                }
            }
        },
        "http.MasTransaction": {
            "type": "object",
            "properties": {
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TransactionTransfer"
                    }
                }
            }
        },
        "http.NilBody": {
            "type": "object"
        },
...
        "http.Response-http_MasTransaction": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/http.MasTransaction"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "http.Response-http_NilBody": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/http.NilBody"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "http.Response-http_TransactionCount": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/http.TransactionCount"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "http.Response-http_TransactionCreateResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/http.TransactionCreateResponse"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "http.Response-string": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "http.ResponseError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "http.TransactionCount": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                }
            }
        },
        "http.TransactionCreateResponse": {
            "type": "object",
            "properties": {
                "transaction_id": {
                    "type": "string"
                }
            }
        },
        "http.UpdTransaction": {
            "type": "object",
            "properties": {
                "account_income": {
                    "type": "string"
                },
                "account_outcome": {
                    "type": "string"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "income": {
                    "type": "number"
                },
                "outcome": {
                    "type": "number"
                },
                "payer": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                }
            }
        },
...
        "models.TransactionTransfer": {
            "type": "object",
            "properties": {
                "account_income": {
                    "type": "string"
                },
                "account_outcome": {
                    "type": "string"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "income": {
                    "type": "number"
                },
                "outcome": {
                    "type": "number"
                },
                "payer": {
                    "type": "string"
                }
            }
        },
...
    }
}